+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.18
+
+### Patch Changes
+
+- Updated dependencies [46f46e4]
+  - @ai-sdk/provider-utils@4.0.7
+  - @ai-sdk/anthropic@3.0.14
+
 ## 4.0.17
 
 ### Patch Changes