+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/vue
 
+## 3.0.37
+
+### Patch Changes
+
+- Updated dependencies [b5dab9b]
+  - ai@6.0.37
+
+## 3.0.36
+
+### Patch Changes
+
+- Updated dependencies [46f46e4]
+  - @ai-sdk/provider-utils@4.0.7
+  - ai@6.0.36
+
+## 3.0.35
+
+### Patch Changes
+
+- Updated dependencies [d7e7f1f]
+  - ai@6.0.35
+
 ## 3.0.34
 
 ### Patch Changes